feat: accept direct file upload for webhook/http (s3)
